Add ‘concept_export_value’ to mapping file. Code cleanup and refactoring.
