+++ arches/app/utils/data_management/resource_graphs/exporter.py
@@ -107,6 +107,8 @@ def create_mapping_configuration_file(graphid, data_dir=None):
             export_node['arches_nodeid'] = str(node.nodeid)
             export_node['arches_node_name'] = node.name
             export_node['file_field_name'] = ""
+            if node.datatype in ['concept', 'concept-list', 'domain-value', 'domain-value-list']:
+                export_node['concept_export_value'] = "label"
             export_node['value_type'] = ""
             export_node['data_length'] = ""
             export_node['data_type'] = ""
