+++ arches/app/utils/data_management/resources/formats/csvfile.py
@@ -66,22 +66,46 @@ class CsvWriter(Writer):
     #         csvwriter.writerow({k:v.encode('utf8') for k,v in csv_record.items()})
     #     return csvs_for_export
 
-    def transform_value_for_export(self, datatype, value):
+    def transform_value_for_export(self, datatype, value, domains, concept_export_value):
+
+        def get_concept_export_value(value, domains, concept_export_value):
+            if concept_export_value != None:
+                if concept_export_value == "label" or concept_export_value == "both":
+                    for domain in domains:
+                        if domain['valueid'] == value:
+                            if concept_export_value == "label":
+                                value = domain['label']
+                            elif concept_export_value == "both":
+                                value = value + '|' + domain['label']
+            return value
+
+
         if datatype == 'geojson-feature-collection':
             wkt_geoms = []
             for feature in value['features']:
                 wkt_geoms.append(GEOSGeometry(json.dumps(feature['geometry'])))
             value = GeometryCollection(wkt_geoms)
+        elif datatype in ['concept-list', 'domain-value-list']:
+            new_values = []
+            for val in value:
+                new_val = get_concept_export_value(val, domains, concept_export_value)
+                new_values.append(new_val)
+            value = ','.join(new_values)
+        elif datatype in ['concept', 'domain-value']:
+            value = get_concept_export_value(value, domains, concept_export_value)
         return value
 
     def write_resources(self, resources, resource_export_configs=None):
         csv_records = []
         other_group_records = []
         mapping = {}
+        concept_export_value_lookup = {}
         for node in resource_export_configs['nodes']:
-            mapping[node['arches_nodeid']] = node['file_field_name']
-        csv_header = mapping.values()
-        csv_header.append('ResourceID')
+            if node['file_field_name'] != '':
+                mapping[node['arches_nodeid']] = node['file_field_name']
+            if 'concept_export_value' in node:
+                concept_export_value_lookup[node['arches_nodeid']] = node['concept_export_value']
+        csv_header = ['ResourceID'] + mapping.values()
         csvs_for_export = []
 
         for resource in resources:
@@ -98,7 +122,10 @@ class CsvWriter(Writer):
                     for k in tile['data'].keys():
                             if tile['data'][k] != '' and k in mapping:
                                 if mapping[k] not in csv_record:
-                                    value = self.transform_value_for_export(self.node_datatypes[k], tile['data'][k])
+                                    concept_export_value = None
+                                    if k in concept_export_value_lookup:
+                                        concept_export_value = concept_export_value_lookup[k]
+                                    value = self.transform_value_for_export(self.node_datatypes[k], tile['data'][k], resource['_source']['domains'], concept_export_value)
                                     csv_record[mapping[k]] = value
                                     del tile['data'][k]
                                 else:
@@ -115,6 +142,7 @@ class CsvWriter(Writer):
         iso_date = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
         csv_name = os.path.join('{0}_{1}.{2}'.format(csv_name_prefix, iso_date, 'csv'))
         dest = StringIO()
+        dest = open('/Users/ryan/desktop/test_export.csv', 'wb')
         csvwriter = csv.DictWriter(dest, delimiter=',', fieldnames=csv_header)
         csvwriter.writeheader()
         csvs_for_export.append({'name':csv_name, 'outputfile': dest})
